+++ System/include/pwm.h
@@ -15,20 +15,20 @@
  *         dutyCycle sets the PWM signal duty cycle.
  *@return: none
 */
-void pwm_init(uint16_t period, uint16_t dutyCycle);
+void pwm_init(uint32_t period, uint32_t dutyCycle);
 
 /** ------------set_left_duty_cycle------------
  *@brief:  set the PWM duty cycle for the left motor
  *@params: dutyCycle
  *@return: none
  */
-void set_left_duty_cycle(uint16_t leftDutyCycle);
+void set_left_duty_cycle(uint32_t leftDutyCycle);
 
 /** ------------set_right_duty_cycle------------
  *@brief:  set the PWM duty cycle for the right motor
  *@params: dutyCycle
  *@return: none
  */
-void set_right_duty_cycle(uint16_t rightDutyCycle);
+void set_right_duty_cycle(uint32_t rightDutyCycle);
 
 #endif /* SYSTEM_INCLUDE_PWM_H_ */
